Diagnosing the identification of the model

Estimate parameters using small posterior samples from JAGS, calculate the approximate Fisher information matrix and return the condition number and rank of the approximate Fisher information matrix with asymptotic variances of estimated parameters.

diagnose_identification(jags_code, data, variable.names, ll_fun, data_fun)

Arguments

jags_code

string. JAGS model code

data

list. Data list of either vector or number

variable.names

vector of string. Names of the variables

ll_fun

function. Log-likelihood function that takes two arguments - data and theta (parameters)

data_fun

function. Function that generates the dataset using theta (parameters). argument of this function is a list of arguments required for creating the dataset.

Value

list. A list of the condition number, rank, and asymptotic variances.

Examples

ll_cross <- function(data, theta){
se = theta[1]
sp = theta[2]
pi = theta[3:5]
pi = c(pi, 1-sum(pi))
eta =  c(se*pi[1]+(1-sp)*pi[3],
  se*pi[2]+(1-sp)*pi[4],
  sp*pi[3]+(1-se)*pi[1],
  (1-se)*pi[2]+sp*pi[4])
  log(dmultinom(data, prob = eta))
  }

  data_cross <- function(args) {
    theta = args$theta
    n = args$n
    se = theta[1]
    sp = theta[2]
    pi = theta[3:5]
    pi = c(pi, 1-sum(pi))
    eta =  c(se*pi[1]+(1-sp)*pi[3],
      se*pi[2]+(1-sp)*pi[4],
      sp*pi[3]+(1-se)*pi[1],
      (1-se)*pi[2]+sp*pi[4])
      as.vector(rmultinom(1, size = n, prob = eta))
     }
jags_code <-
"model{
y[1:4] ~ dmulti(eta[1:4], n)
eta[1:4] <- c(se*pi[1]+(1-sp)*pi[3],
se*pi[2]+(1-sp)*pi[4],
sp*pi[3]+(1-se)*pi[1],
(1-se)*pi[2]+sp*pi[4])
pi[1:4] ~ ddirch(c(1,1,1,1))
se ~ dbeta(25,3)
sp ~ dbeta(30,1.5)
}"
data(small_cross)
data = list(y = small_cross, n = sum(small_cross))
diagnose_identification(jags_code, data, c("se","sp","pi[1]","pi[2]","pi[3]"),ll_cross,data_cross)
#>   |                                                          |                                                  |   0%  |                                                          |*                                                 |   2%  |                                                          |**                                                |   4%  |                                                          |***                                               |   6%  |                                                          |****                                              |   8%  |                                                          |*****                                             |  10%  |                                                          |******                                            |  12%  |                                                          |*******                                           |  14%  |                                                          |********                                          |  16%  |                                                          |*********                                         |  18%  |                                                          |**********                                        |  20%  |                                                          |***********                                       |  22%  |                                                          |************                                      |  24%  |                                                          |*************                                     |  26%  |                                                          |**************                                    |  28%  |                                                          |***************                                   |  30%  |                                                          |****************                                  |  32%  |                                                          |*****************                                 |  34%  |                                                          |******************                                |  36%  |                                                          |*******************                               |  38%  |                                                          |********************                              |  40%  |                                                          |*********************                             |  42%  |                                                          |**********************                            |  44%  |                                                          |***********************                           |  46%  |                                                          |************************                          |  48%  |                                                          |*************************                         |  50%  |                                                          |**************************                        |  52%  |                                                          |***************************                       |  54%  |                                                          |****************************                      |  56%  |                                                          |*****************************                     |  58%  |                                                          |******************************                    |  60%  |                                                          |*******************************                   |  62%  |                                                          |********************************                  |  64%  |                                                          |*********************************                 |  66%  |                                                          |**********************************                |  68%  |                                                          |***********************************               |  70%  |                                                          |************************************              |  72%  |                                                          |*************************************             |  74%  |                                                          |**************************************            |  76%  |                                                          |***************************************           |  78%  |                                                          |****************************************          |  80%  |                                                          |*****************************************         |  82%  |                                                          |******************************************        |  84%  |                                                          |*******************************************       |  86%  |                                                          |********************************************      |  88%  |                                                          |*********************************************     |  90%  |                                                          |**********************************************    |  92%  |                                                          |***********************************************   |  94%  |                                                          |************************************************  |  96%  |                                                          |************************************************* |  98%  |                                                          |**************************************************| 100%
#>   |                                                          |                                                  |   0%  |                                                          |*                                                 |   2%  |                                                          |**                                                |   4%  |                                                          |***                                               |   6%  |                                                          |****                                              |   8%  |                                                          |*****                                             |  10%  |                                                          |******                                            |  12%  |                                                          |*******                                           |  14%  |                                                          |********                                          |  16%  |                                                          |*********                                         |  18%  |                                                          |**********                                        |  20%  |                                                          |***********                                       |  22%  |                                                          |************                                      |  24%  |                                                          |*************                                     |  26%  |                                                          |**************                                    |  28%  |                                                          |***************                                   |  30%  |                                                          |****************                                  |  32%  |                                                          |*****************                                 |  34%  |                                                          |******************                                |  36%  |                                                          |*******************                               |  38%  |                                                          |********************                              |  40%  |                                                          |*********************                             |  42%  |                                                          |**********************                            |  44%  |                                                          |***********************                           |  46%  |                                                          |************************                          |  48%  |                                                          |*************************                         |  50%  |                                                          |**************************                        |  52%  |                                                          |***************************                       |  54%  |                                                          |****************************                      |  56%  |                                                          |*****************************                     |  58%  |                                                          |******************************                    |  60%  |                                                          |*******************************                   |  62%  |                                                          |********************************                  |  64%  |                                                          |*********************************                 |  66%  |                                                          |**********************************                |  68%  |                                                          |***********************************               |  70%  |                                                          |************************************              |  72%  |                                                          |*************************************             |  74%  |                                                          |**************************************            |  76%  |                                                          |***************************************           |  78%  |                                                          |****************************************          |  80%  |                                                          |*****************************************         |  82%  |                                                          |******************************************        |  84%  |                                                          |*******************************************       |  86%  |                                                          |********************************************      |  88%  |                                                          |*********************************************     |  90%  |                                                          |**********************************************    |  92%  |                                                          |***********************************************   |  94%  |                                                          |************************************************  |  96%  |                                                          |************************************************* |  98%  |                                                          |**************************************************| 100%
#> $condition_number
#> [1] 3528.201
#> 
#> $rank
#> [1] 5
#> 
#> $var
#> [1] 5.5016237 0.2782333 0.3289914 0.8136337 0.2927017
#>