Posterior sampling of given JAGS (BUGS) code and data

This function returns the results of sampling from posteriors of given model and data using either importance sampling with transparent reparameterization (ISTP) or pigeons algorithm.

sampling(
  jags_code,
  data,
  type = c("sat", "cross", "pigeons"),
  extract_variable_names = NULL,
  n_chains = NULL,
  n_rounds = NULL,
  N = NULL,
  extended_traces = F
)

Arguments

jags_code

A string of JAGS code that specifies the model.

data

A list of dataset.

type

one of "sat", "cross", "pigeons"

extract_variable_names

(optional) A vector of strings that contain the variable names that will be extracted

n_chains

n_chains for Pigeons.jl The size of final sample is n_chains * 2^(n_rounds)

n_rounds

n_rounds for Pigeons.jl The size of final sample is n_chains * 2^(n_rounds)

N

type sat and cross will generate the posterior sample of size N

extended_traces

logical. If TRUE, all serialized chains are recorded.

Value

Either a list or julia object. Results including samples.

Examples

jags_code <-
"model{
y[1:4] ~ dmulti(eta[1:4], n)
eta[1:4] <- c(se*pi[1]+(1-sp)*pi[3],
se*pi[2]+(1-sp)*pi[4],
sp*pi[3]+(1-se)*pi[1],
(1-se)*pi[2]+sp*pi[4])
pi[1:4] ~ ddirch(c(1,1,1,1))
se ~ dbeta(25,3)
sp ~ dbeta(30,1.5)
}"
data(small_cross)
data = list(y = small_cross, n = sum(small_cross))
results = sampling(jags_code, data, "cross", N = 15000)
#> Julia version 1.10.7 at location /Applications/Julia-1.10.app/Contents/Resources/julia/bin will be used.
#> Loading setup script for JuliaCall...
#> Finish loading setup script for JuliaCall.
#> Int64wrapper (generic function with 1 method)
#>   
#> c (generic function with 1 method)
#>   
#> (y = [7, 13, 6, 12], n = 38)
#>   
#> Any[]
#>   
#> (y = [7, 13, 6, 12], n = 38)
#>   
#> quote
#>     y[1:4] ~ dmulti(eta[1:4], n)
#>     eta[1:4] = c(se * pi[1] + (1 - sp) * pi[3], se * pi[2] + (1 - sp) * pi[4], sp * pi[3] + (1 - se) * pi[1], (1 - se) * pi[2] + sp * pi[4])
#>     pi[1:4] ~ ddirich(c(1, 1, 1, 1))
#>     se ~ dbeta(25, 3)
#>     sp ~ dbeta(30, 1.5)
#> end
#>   
#> BUGSModel (parameters are in transformed (unconstrained) space, with dimension 5):
#> 
#>   Model parameters:
#>     se
#>     pi[1:4]
#>     sp
#> 
#>   Variable sizes and types:
#>     n: type = Int64
#>     y: size = (4,), type = Vector{Int64}
#>     eta: size = (4,), type = Vector{Float64}
#>     se: type = Float64
#>     pi: size = (4,), type = Vector{Float64}
#>     sp: type = Float64
#> 
#>   
#> 3-element Vector{AbstractPPL.VarName}:
#>  sp
#>  se
#>  pi[1:4]
#>   
#> JuliaBUGS.FlattenedGraphNodeData{Vector{Function}, Vector{@NamedTuple{}}}(AbstractPPL.VarName[sp, se, pi[1:4], eta[1:4], y[1:4]], Bool[1, 1, 1, 0, 1], Bool[0, 0, 0, 0, 1], Function[JuliaBUGS.var"#56#57"(), JuliaBUGS.var"#54#55"(), JuliaBUGS.var"#52#53"(), JuliaBUGS.var"#50#51"(), JuliaBUGS.var"#48#49"()], [NamedTuple(), NamedTuple(), NamedTuple(), NamedTuple(), NamedTuple()])
#>   
#> 5-element Vector{AbstractPPL.VarName}:
#>  sp
#>  se
#>  pi[1:4]
#>  eta[1:4]
#>  y[1:4]
#>   
#> (n = 38, y = [7, 13, 6, 12], eta = [0.14584394794661662, 0.26151299458887156, 0.20183229100792474, 0.3908107664565872], se = 0.8173923326894553, pi = [0.16425516595554243, 0.29210606651306675, 0.18342107299899893, 0.36021769453239194], sp = 0.9368500330930601)
#>   
#> String[]
#>   
#> Bool[]
#>   
#> Bool[]
#>   
results$ess
#> [1] 14437.12
head(results$samples)
#>          pi[11]    pi[12]     pi[21]        se        sp
#> [1,] 0.20124085 0.2619542 0.23260845 0.8520560 0.9785705
#> [2,] 0.17275585 0.5495682 0.11387995 0.7586394 0.9496084
#> [3,] 0.18722017 0.2542498 0.22603068 0.9252069 0.9883285
#> [4,] 0.19349624 0.2576683 0.18487586 0.9443064 0.8996383
#> [5,] 0.22252634 0.2956584 0.09739528 0.8709533 0.9933115
#> [6,] 0.09628142 0.4256138 0.24565598 0.8877216 0.8966986

data(small_sat)
sigma = 0.5 #Hyperparameter, prior variance of log odds ratios
data = c(small_sat, list(n = length(small_sat$r),
  p2 = dim(small_sat$x)[2],
  tau = 1/sigma^2))
jags_code  <- "model{
  for (i in 1:n) {
    x[i, 1:p2] ~ dmulti(px[1:p2],1)
    y[i] ~ dbern(inprod(x[i,1:p2], py_x[1:p2]))
    r[i] ~ dbern(pr[i])
    pr[i] <- (y[i]) * inprod(x[i,1:p2], pr_xy1[1:p2]) + abs(y[i]-1) * inprod(x[i,1:p2], pr_xy0[1:p2])
  }
  for(i in 1:p2){
    pr_xy1[i] <- ilogit(log_or_r[i] + logit(pr_xy0[i]))
  }
  px[1:p2] ~ ddirich(Int64wrapper(ones(p2)))
  for(i in 1:p2) {
    py_x[i] ~ dunif(0, 1)
    pr_xy0[i] ~ dunif(0, 1)
    log_or_r[i] ~ dnorm(0, tau)
  }
}"
results = sampling(jags_code, data, "sat", N = 15000)
#> Int64wrapper (generic function with 1 method)
#>   
#> c (generic function with 1 method)
#>   
#> (x = [1 0 0 0; 0 0 1 0; … ; 0 0 1 0; 0 0 1 0], y = Union{Missing, Int64}[missing, missing, 1, 1, missing, 1, 1, 0, missing, 0  …  missing, missing, 1, missing, missing, 1, missing, 0, missing, missing], r = [0.0, 0.0, 1.0, 1.0, 0.0, 1.0, 1.0, 1.0, 0.0, 1.0  …  0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 1.0, 0.0, 0.0], n = 100, p2 = 4, tau = 4.0)
#>   
#> Any[]
#>   
#> (x = [1 0 0 0; 0 0 1 0; … ; 0 0 1 0; 0 0 1 0], y = Union{Missing, Int64}[missing, missing, 1, 1, missing, 1, 1, 0, missing, 0  …  missing, missing, 1, missing, missing, 1, missing, 0, missing, missing], r = [0, 0, 1, 1, 0, 1, 1, 1, 0, 1  …  0, 0, 1, 0, 0, 1, 0, 1, 0, 0], n = 100, p2 = 4, tau = 4)
#>   
#> quote
#>     for i = 1:n
#>         x[i, 1:p2] ~ dmulti(px[1:p2], 1)
#>         y[i] ~ dbern(inprod(x[i, 1:p2], py_x[1:p2]))
#>         r[i] ~ dbern(pr[i])
#>         pr[i] = y[i] * inprod(x[i, 1:p2], pr_xy1[1:p2]) + abs(y[i] - 1) * inprod(x[i, 1:p2], pr_xy0[1:p2])
#>     end
#>     for i = 1:p2
#>         pr_xy1[i] = ilogit(log_or_r[i] + logit(pr_xy0[i]))
#>     end
#>     px[1:p2] ~ ddirich(Int64wrapper(ones(p2)))
#>     for i = 1:p2
#>         py_x[i] ~ dunif(0, 1)
#>         pr_xy0[i] ~ dunif(0, 1)
#>         log_or_r[i] ~ dnorm(0, tau)
#>     end
#> end
#>   
#> BUGSModel (parameters are in transformed (unconstrained) space, with dimension 68):
#> 
#>   Model parameters:
#>     y[100], y[99], y[97], y[95], y[94], y[92], y[91], y[90], y[87], y[86], y[82], y[79], y[75], y[74], y[72], y[71], y[70], y[67], y[62], y[61], y[57], y[56], y[55], y[54], y[52], y[48], y[47], y[45], y[44], y[43], y[41], y[40], y[37], y[36], y[34], y[33], y[29], y[28], y[27], y[24], y[22], y[21], y[19], y[17], y[15], y[14], y[13], y[12], y[11], y[9], y[5], y[2], y[1]
#>     pr_xy0[4], pr_xy0[3], pr_xy0[2], pr_xy0[1]
#>     px[1:4]
#>     log_or_r[4], log_or_r[3], log_or_r[2], log_or_r[1]
#>     py_x[4], py_x[3], py_x[2], py_x[1]
#> 
#>   Variable sizes and types:
#>     n: type = Int64
#>     pr_xy0: size = (4,), type = Vector{Float64}
#>     pr: size = (100,), type = Vector{Float64}
#>     px: size = (4,), type = Vector{Float64}
#>     x: size = (100, 4), type = Matrix{Int64}
#>     p2: type = Int64
#>     pr_xy1: size = (4,), type = Vector{Float64}
#>     tau: type = Int64
#>     r: size = (100,), type = Vector{Int64}
#>     log_or_r: size = (4,), type = Vector{Float64}
#>     y: size = (100,), type = Vector{Int64}
#>     py_x: size = (4,), type = Vector{Float64}
#> 
#>   
#> 66-element Vector{AbstractPPL.VarName}:
#>  log_or_r[4]
#>  pr_xy0[4]
#>  py_x[4]
#>  log_or_r[3]
#>  pr_xy0[3]
#>  py_x[3]
#>  log_or_r[2]
#>  pr_xy0[2]
#>  py_x[2]
#>  log_or_r[1]
#>  ⋮
#>  y[15]
#>  y[14]
#>  y[13]
#>  y[12]
#>  y[11]
#>  y[9]
#>  y[5]
#>  y[2]
#>  y[1]
#>   
#> JuliaBUGS.FlattenedGraphNodeData{Vector{Function}, Vector{NamedTuple}}(AbstractPPL.VarName[log_or_r[4], pr_xy0[4], py_x[4], log_or_r[3], pr_xy0[3], py_x[3], log_or_r[2], pr_xy0[2], py_x[2], log_or_r[1]  …  y[3], x[3, 1:4], y[2], pr[2], r[2], x[2, 1:4], y[1], pr[1], r[1], x[1, 1:4]], Bool[1, 1, 1, 1, 1, 1, 1, 1, 1, 1  …  1, 1, 1, 0, 1, 1, 1, 0, 1, 1], Bool[0, 0, 0, 0, 0, 0, 0, 0, 0, 0  …  1, 1, 0, 0, 1, 1, 0, 0, 1, 1], Function[JuliaBUGS.var"#74#75"(), JuliaBUGS.var"#72#73"(), JuliaBUGS.var"#70#71"(), JuliaBUGS.var"#74#75"(), JuliaBUGS.var"#72#73"(), JuliaBUGS.var"#70#71"(), JuliaBUGS.var"#74#75"(), JuliaBUGS.var"#72#73"(), JuliaBUGS.var"#70#71"(), JuliaBUGS.var"#74#75"()  …  JuliaBUGS.var"#60#61"(), JuliaBUGS.var"#58#59"(), JuliaBUGS.var"#60#61"(), JuliaBUGS.var"#64#65"(), JuliaBUGS.var"#62#63"(), JuliaBUGS.var"#58#59"(), JuliaBUGS.var"#60#61"(), JuliaBUGS.var"#64#65"(), JuliaBUGS.var"#62#63"(), JuliaBUGS.var"#58#59"()], NamedTuple[(i = 4,), (i = 4,), (i = 4,), (i = 3,), (i = 3,), (i = 3,), (i = 2,), (i = 2,), (i = 2,), (i = 1,)  …  (i = 3,), (i = 3,), (i = 2,), (i = 2,), (i = 2,), (i = 2,), (i = 1,), (i = 1,), (i = 1,), (i = 1,)])
#>   
#> 417-element Vector{AbstractPPL.VarName}:
#>  log_or_r[4]
#>  pr_xy0[4]
#>  py_x[4]
#>  log_or_r[3]
#>  pr_xy0[3]
#>  py_x[3]
#>  log_or_r[2]
#>  pr_xy0[2]
#>  py_x[2]
#>  log_or_r[1]
#>  ⋮
#>  x[3, 1:4]
#>  y[2]
#>  pr[2]
#>  r[2]
#>  x[2, 1:4]
#>  y[1]
#>  pr[1]
#>  r[1]
#>  x[1, 1:4]
#>   
#> (n = 100, pr_xy0 = [0.07841117781060947, 0.8144951039775508, 0.6333081377267009, 0.3248283777658638], pr = [0.07841117781060947, 0.6325040621753106, 0.6325040621753106, 0.6325040621753106, 0.6325040621753106, 0.2761007965129682, 0.05913329118663244, 0.07841117781060947, 0.07841117781060947, 0.6333081377267009  …  0.7980364147261364, 0.6325040621753106, 0.2761007965129682, 0.6325040621753106, 0.7980364147261364, 0.2761007965129682, 0.7980364147261364, 0.8144951039775508, 0.6333081377267009, 0.6325040621753106], px = [0.4330045288844766, 0.1565482705882617, 0.13062596662102074, 0.27982123390624103], x = [1 0 0 0; 0 0 1 0; … ; 0 0 1 0; 0 0 1 0], p2 = 4, pr_xy1 = [0.05913329118663244, 0.7980364147261364, 0.6325040621753106, 0.2761007965129682], tau = 4, r = [0, 0, 1, 1, 0, 1, 1, 1, 0, 1  …  0, 0, 1, 0, 0, 1, 0, 1, 0, 0], log_or_r = [-0.3028747427816763, -0.10542032292695717, -0.0034608321708376696, -0.2322162179101596], y = [0, 1, 1, 1, 1, 1, 1, 0, 0, 0  …  1, 1, 1, 1, 1, 1, 1, 0, 0, 1], py_x = [0.27400861506520446, 0.5160997847486777, 0.7796847802957392, 0.2251398944266806])
#>   
#> String[]
#>   
#> Bool[]
#>   
#> Bool[]
#>   

results$ess
#> [1] 10553.91
head(results$samples)
#>           px[2]     px[3]     px[4]    py_x[1]     py_x[2]   py_x[3]   py_x[4]
#> [1,] 0.04954925 0.4557230 0.2720335 0.20806585 0.008694443 0.2369732 0.6007577
#> [2,] 0.03074049 0.5093056 0.1999754 0.34790710 0.091252438 0.2441334 0.4191591
#> [3,] 0.04348269 0.6167857 0.1986876 0.24955974 0.022393921 0.2160827 0.4495713
#> [4,] 0.04318535 0.4709800 0.2312873 0.32393568 0.020220343 0.2899093 0.3660454
#> [5,] 0.02233342 0.5357085 0.2741371 0.04213234 0.014291204 0.2713262 0.7402509
#> [6,] 0.07432488 0.5123498 0.1808182 0.29612516 0.063756676 0.1911955 0.4520687
#>      pr_xy0[1] pr_xy0[2] pr_xy0[3] pr_xy0[4] log_or_r[1] log_or_r[2]
#> [1,] 0.3803292 0.3002579 0.2935402 0.9066436 -0.33016649  0.19498591
#> [2,] 0.3543570 0.5550753 0.3975976 0.7808745 -0.25790062 -0.85711089
#> [3,] 0.2429704 0.3680197 0.3980622 0.8007605  0.58416903  0.47163354
#> [4,] 0.4444687 0.3472958 0.3499215 0.9580744 -0.39840020  0.84930739
#> [5,] 0.3948556 0.3131689 0.3848787 0.8140979  0.22989351 -0.07674268
#> [6,] 0.4399092 0.3570549 0.3051515 0.7909865 -0.06501844 -0.70129097
#>      log_or_r[3] log_or_r[4]
#> [1,]  -0.2528078  0.54904463
#> [2,]  -0.2443801 -0.08111126
#> [3,]  -0.4657444  0.46419913
#> [4,]   0.2686651  0.38568747
#> [5,]  -0.5541823 -0.14034874
#> [6,]  -0.1119188  1.27288191

jags_code <- "model{
for (i in 1:n) {
r[i] ~ dbern(pr[i])
pr[i] <- ilogit(y[i] * alpha1 + alpha0)
y[i] ~ dpois(mu)
}
mu ~ dgamma(1,1)
alpha0 ~ dnorm(0, 0.1)
alpha1 ~ dnorm(0, tau)
}
"
data(small_count)
sigma = 0.5 #Hyperparameter, prior variance of alpha1
data = c(small_count, list(n = length(small_count$r), tau = 1/sigma^2))
sampling(jags_code, data, "pigeons", n_chains =  4, n_rounds = 4)
#> Int64wrapper (generic function with 1 method)
#>   
#> c (generic function with 1 method)
#>   
#> (y = Union{Missing, Int64}[6, missing, missing, missing, 1, 3, 1, 5, 1, missing  …  missing, missing, missing, 1, 3, missing, missing, 1, 3, missing], r = [1, 0, 0, 0, 1, 1, 1, 1, 1, 0  …  0, 0, 0, 1, 1, 0, 0, 1, 1, 0], n = 100, tau = 4.0)
#>   
#> Any[]
#>   
#> (y = Union{Missing, Int64}[6, missing, missing, missing, 1, 3, 1, 5, 1, missing  …  missing, missing, missing, 1, 3, missing, missing, 1, 3, missing], r = [1, 0, 0, 0, 1, 1, 1, 1, 1, 0  …  0, 0, 0, 1, 1, 0, 0, 1, 1, 0], n = 100, tau = 4)
#>   
#> quote
#>     for i = 1:n
#>         r[i] ~ dbern(pr[i])
#>         pr[i] = ilogit(y[i] * alpha1 + alpha0)
#>         y[i] ~ dpois(mu)
#>     end
#>     mu ~ dgamma(1, 1)
#>     alpha0 ~ dnorm(0, 0.1)
#>     alpha1 ~ dnorm(0, tau)
#> end
#>   
#> BUGSModel (parameters are in transformed (unconstrained) space, with dimension 47):
#> 
#>   Model parameters:
#>     mu
#>     y[100], y[97], y[96], y[93], y[92], y[91], y[90], y[87], y[84], y[83], y[81], y[80], y[78], y[77], y[75], y[71], y[70], y[69], y[67], y[64], y[57], y[54], y[52], y[51], y[50], y[48], y[47], y[42], y[41], y[39], y[38], y[36], y[34], y[31], y[30], y[28], y[26], y[22], y[16], y[12], y[10], y[4], y[3], y[2]
#>     alpha1
#>     alpha0
#> 
#>   Variable sizes and types:
#>     n: type = Int64
#>     y: size = (100,), type = Vector{Int64}
#>     mu: type = Float64
#>     pr: size = (100,), type = Vector{Float64}
#>     tau: type = Int64
#>     r: size = (100,), type = Vector{Int64}
#>     alpha0: type = Float64
#>     alpha1: type = Float64
#> 
#>   
#> JuliaBUGSPath{...}(BUGSModel (parameters are in transformed (unconstrained) space, with dimension 47):
#> 
#>   Model parameters:
#>     mu
#>     y[100], y[97], y[96], y[93], y[92], y[91], y[90], y[87], y[84], y[83], y[81], y[80], y[78], y[77], y[75], y[71], y[70], y[69], y[67], y[64], y[57], y[54], y[52], y[51], y[50], y[48], y[47], y[42], y[41], y[39], y[38], y[36], y[34], y[31], y[30], y[28], y[26], y[22], y[16], y[12], y[10], y[4], y[3], y[2]
#>     alpha1
#>     alpha0
#> 
#>   Variable sizes and types:
#>     n: type = Int64
#>     y: size = (100,), type = Vector{Int64}
#>     mu: type = Float64
#>     pr: size = (100,), type = Vector{Float64}
#>     tau: type = Int64
#>     r: size = (100,), type = Vector{Int64}
#>     alpha0: type = Float64
#>     alpha1: type = Float64
#> )
#>   
#> PT(checkpoint = false, ...)
#>   
#> Julia Object of type PT{...}.
#> PT(checkpoint = false, ...)